:sparkles: feat: Validações referentes à autenticação na entidade consulta

+++ src/main/java/com/connectCare/connectCareApi/services/impl/ConsultaServiceImpl.java
@@ -2,6 +2,7 @@ package com.connectCare.connectCareApi.services.impl;
 
 import com.connectCare.connectCareApi.exceptions.ConsultaNaoEncontradaException;
 import com.connectCare.connectCareApi.exceptions.HorarioAgendadoException;
+import com.connectCare.connectCareApi.exceptions.NaoAutorizadoException;
 import com.connectCare.connectCareApi.exceptions.NenhumRegistroEncontradoException;
 import com.connectCare.connectCareApi.exceptions.OperacaoBancoDeDadosException;
 import com.connectCare.connectCareApi.models.dtos.GetPorIntervaloDataDTO;
@@ -11,13 +12,17 @@ import com.connectCare.connectCareApi.models.entities.Medico;
 import com.connectCare.connectCareApi.models.entities.Paciente;
 import com.connectCare.connectCareApi.repositories.ConsultaRepository;
 import com.connectCare.connectCareApi.services.GenericService;
+import com.connectCare.connectCareApi.utils.UsuarioAutenticado;
+
 import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 
 @Service
 public class ConsultaServiceImpl implements GenericService<Consulta> {
@@ -47,6 +52,11 @@ public class ConsultaServiceImpl implements GenericService<Consulta> {
 
         Medico medicoEncontrado = medicoService.getById(consulta.getMedico().getId());
         consulta.setMedico(medicoEncontrado);
+        
+        //Verificando se o ID do usuário autenticado é igual ao ID do usuário associado ao paciente
+		if(!Objects.equals(pacienteEncontrado.getUsuario().getId(), UsuarioAutenticado.getUsuarioAutenticado().getId())){
+			throw new NaoAutorizadoException();
+		}
 
         Disponibilidade disponibilidadeEncontrada = disponibilidadeService.getById(consulta.getDisponibilidade().getId());
 
@@ -70,13 +80,28 @@ public class ConsultaServiceImpl implements GenericService<Consulta> {
 
     public List<Consulta> getByPacienteId(Integer id){
         List<Consulta> consultasEncontradas = repository.findByPacienteId(id);
-
+        
+        Paciente pacienteEncontrado = pacienteService.getById(id);
+        
+        //Verificando se o ID do usuário autenticado é igual ao ID do usuário associado ao paciente
+        if(!Objects.equals(pacienteEncontrado.getUsuario().getId(), UsuarioAutenticado.getUsuarioAutenticado().getId())){
+            throw new NaoAutorizadoException();
+        }
+        
         if(consultasEncontradas.isEmpty()) throw new NenhumRegistroEncontradoException("Consulta");
         return consultasEncontradas;
     }
 
     public List<Consulta> getByMedicoId(Integer id){
         List<Consulta> consultasEncontradas = repository.findByMedicoId(id);
+        
+        Medico medicoEncontrado = medicoService.getById(id);
+        
+        //Verificando se o ID do usuário autenticado é igual ao ID do usuário associado ao medico
+        if(!Objects.equals(medicoEncontrado.getUsuario().getId(), UsuarioAutenticado.getUsuarioAutenticado().getId())){
+            throw new NaoAutorizadoException();
+        }
+        
 
         if(consultasEncontradas.isEmpty()) throw new NenhumRegistroEncontradoException("Consulta");
         return consultasEncontradas;
@@ -85,6 +110,13 @@ public class ConsultaServiceImpl implements GenericService<Consulta> {
     public List<Consulta> getByPacienteIdData(Integer id, GetPorIntervaloDataDTO intervaloData){
         List<Consulta> consultasEncontradas = repository.findByPacienteIdAndDisponibilidadeDataDisponivelBetween(id, intervaloData.getInicio(), intervaloData.getFim());
 
+        Paciente pacienteEncontrado = pacienteService.getById(id);
+        
+        //Verificando se o ID do usuário autenticado é igual ao ID do usuário associado ao paciente
+        if(!Objects.equals(pacienteEncontrado.getUsuario().getId(), UsuarioAutenticado.getUsuarioAutenticado().getId())){
+            throw new NaoAutorizadoException();
+        }
+        
         if(consultasEncontradas.isEmpty()) throw new NenhumRegistroEncontradoException("Consulta");
         return consultasEncontradas;
     }
@@ -92,6 +124,11 @@ public class ConsultaServiceImpl implements GenericService<Consulta> {
     @Override
     public List<Consulta> getAll() {
         List<Consulta> consultasEncontradas = repository.findAll();
+        
+        //Verificando se o atual usuário é ADMIN
+        if(!UsuarioAutenticado.getUsuarioAutenticado().getAuthorities().contains(new SimpleGrantedAuthority("ADMIN"))){
+            throw new NaoAutorizadoException();
+        }
 
         if(consultasEncontradas.isEmpty()) throw new NenhumRegistroEncontradoException("Consulta");
         return consultasEncontradas;
@@ -101,6 +138,13 @@ public class ConsultaServiceImpl implements GenericService<Consulta> {
     public Consulta update(Consulta consulta) {
         try{
             Consulta consultaEncontrada = repository.getReferenceById(consulta.getId());
+            
+            Paciente pacienteEncontrado = consultaEncontrada.getPaciente();
+            
+            //Verificando se o ID do usuário autenticado é igual ao ID do usuário associado ao paciente
+    		if(!Objects.equals(pacienteEncontrado.getUsuario().getId(), UsuarioAutenticado.getUsuarioAutenticado().getId())){
+    			throw new NaoAutorizadoException();
+    		}
 
             consultaEncontrada.setTipoConsulta(consulta.getTipoConsulta());
             consultaEncontrada.setFormaAgendamento(consulta.getFormaAgendamento());
@@ -117,6 +161,14 @@ public class ConsultaServiceImpl implements GenericService<Consulta> {
     public void delete(Integer id) {
         try{
             Consulta consultaEncontrada = repository.findById(id).orElseThrow(() -> new ConsultaNaoEncontradaException(id));
+            
+            Paciente pacienteEncontrado = consultaEncontrada.getPaciente();
+            
+            //Verificando se o ID do usuário autenticado é igual ao ID do usuário associado ao paciente
+    		if(!Objects.equals(pacienteEncontrado.getUsuario().getId(), UsuarioAutenticado.getUsuarioAutenticado().getId())){
+    			throw new NaoAutorizadoException();
+    		}
+    		
             repository.delete(consultaEncontrada);
         }catch(DataIntegrityViolationException e){
             throw new OperacaoBancoDeDadosException();
